Check fieldmask in case of network-ipam ref update
In case of virtual network update check if field mask is set for
network-ipam refs.

Partial-bug: #1776234

+++ pkg/db/address_manager.go
@@ -74,6 +74,15 @@ func (db *Service) DeleteIpamSubnet(ctx context.Context, request *ipam.DeleteIpa
 		return errors.Errorf("empty subnet uuid in DeleteIpamSubnet")
 	}
 
+	subnetExists, err := db.CheckIfIpamSubnetExists(ctx, request.SubnetUUID)
+	if err != nil {
+		return err
+	}
+
+	if !subnetExists {
+		return errors.Errorf("ipam subnet with uuid %s doesn't exist", request.SubnetUUID)
+	}
+
 	return db.deleteIPPools(ctx, &ipPool{
 		key: request.SubnetUUID,
 	})
